uno/requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
